Buildathon-main/requirements.txt
# Broadway Show & Event Planner Requirements
# Application dependencies organized by category

# Core Dependencies
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1

# API Integrations
openai==1.3.0
googlemaps==4.10.0
google-api-python-client==2.108.0
pytz==2023.3

# Data Processing & ML
numpy==1.24.4
pandas==2.0.3
sentence-transformers==2.2.2
scikit-learn==1.3.0

# Development & Testing
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.1.0

# Web & UI
flask-cors==4.0.0
jinja2==3.1.2
gunicorn==21.2.0  # For production deployment

# Utility
cachetools==5.3.1
python-dateutil==2.8.2
retry==0.9.2 